# api/docs/swagger.yaml
definitions:
...
  models.CreateStaff:
    properties:
      age:
        type: integer
      balance:
        type: integer
      birthdate:
        type: integer
      branch_id:
        type: string
      ctreate_at:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      staff_type_enum:
        type: string
      tarif_id:
        type: string
    type: object
...
  models.CreateStorage:
    properties:
      branch_id:
        type: string
      count:
        type: integer
      ctreate_at:
        type: string
      product_id:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.Staff:
    properties:
      age:
        type: integer
      balance:
        type: integer
      birthdate:
        type: integer
      branch_id:
        type: string
      ctreate_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      staff_type_enum:
        type: string
      tarif_id:
        type: string
    type: object
...
  models.Storage:
    properties:
      branch_id:
        type: string
      count:
        type: integer
      ctreate_at:
        type: string
      id:
        type: string
      product_id:
        type: string
    type: object
...
  models.UpdateStaff:
    properties:
      age:
        type: integer
      balance:
        type: integer
      birthdate:
        type: integer
      branch_id:
        type: string
      ctreate_at:
        type: string
      id:
        type: string
      name:
        type: string
      tarif_id:
        type: string
    type: object
...
  models.UpdateStorage:
    properties:
      branch_id:
        type: string
      count:
        type: integer
      id:
        type: string
      product_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /staff:
    post:
      consumes:
      - application/json
      description: create a new staff
      parameters:
      - description: staff
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new staff
      tags:
      - staff
  /staff/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete staff by ID
      tags:
      - staff
    get:
      consumes:
      - application/json
      parameters:
      - description: Staff ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get staff by ID
      tags:
      - staff
    put:
      consumes:
      - application/json
      description: update staff
      parameters:
      - description: staff
        in: path
        name: id
        required: true
        type: string
      - description: staff
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update staff
      tags:
      - staff
...
  /staffs:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a list of staff
      tags:
      - staff
  /storage:
    post:
      consumes:
      - application/json
      description: create a new storage
      parameters:
      - description: storage
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.CreateStorage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Storage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new storage
      tags:
      - storage
  /storage/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Storage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data successfully deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete storage by ID
      tags:
      - storage
    get:
      consumes:
      - application/json
      parameters:
      - description: Storage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Storage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get storage by ID
      tags:
      - storage
    put:
      consumes:
      - application/json
      description: update storage
      parameters:
      - description: storage
        in: path
        name: id
        required: true
        type: string
      - description: storage
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStorage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Storage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update storage
      tags:
      - storage
  /storages:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Storage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a list of storage
      tags:
      - storage
...
swagger: "2.0"